AJout des modifiers, du makefile et suppresions de leaks

+++ libftprintf/ft_power.c
@@ -6,11 +6,11 @@
 /*   By: rfernand <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/18 14:57:49 by rfernand          #+#    #+#             */
-/*   Updated: 2016/04/02 15:54:43 by rfernand         ###   ########.fr       */
+/*   Updated: 2016/04/20 19:28:49 by rfernand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_power(int i, int power)
+long long		ft_power(long long i, int power)
 {
 	if (power == 0)
 		return (1);
